+++ changelog.md
@@ -19,6 +19,7 @@ __[maintenance]__
 * Update certificate information for MariaDB 10.6 installer.
 * Update certificate information for MariaDB 10.11 installer.
 * Update certificate information for MariaDB 11.4 installer.
+* Update certificate information for MariaDB 11.8 installer.
 * Update certificate information for Node.js installer.
 * Update certificate information for TeamViewer installer.